mov initTable to HoodieTableMetaClient
