+++ hudi-utilities/src/main/java/org/apache/hudi/utilities/HDFSParquetImporter.java
@@ -28,7 +28,6 @@ import org.apache.hudi.common.fs.FSUtils;
 import org.apache.hudi.common.model.HoodieKey;
 import org.apache.hudi.common.model.HoodieRecord;
 import org.apache.hudi.common.model.HoodieRecordPayload;
-import org.apache.hudi.common.table.HoodieTableConfig;
 import org.apache.hudi.common.table.HoodieTableMetaClient;
 import org.apache.hudi.common.util.Option;
 import org.apache.hudi.exception.HoodieIOException;
@@ -135,7 +134,7 @@ public class HDFSParquetImporter implements Serializable {
 
       if (!fs.exists(new Path(cfg.targetPath))) {
         // Initialize target hoodie table.
-        Properties properties = HoodieTableConfig.propertyBuilder()
+        Properties properties = HoodieTableMetaClient.withPropertyBuilder()
             .setTableName(cfg.tableName)
             .setTableType(cfg.tableType)
             .build();
